+++ sentry-sidekiq/spec/spec_helper.rb
@@ -134,7 +134,9 @@ class SadWorker
 end
 
 class HappyWorkerForCron < HappyWorker; end
+class HappyWorkerForCron < HappyWorker; end
 class HappyWorkerForScheduler < HappyWorker; end
+class HappyWorkerForSchedulerWithTimezone < HappyWorker; end
 class EveryHappyWorker < HappyWorker; end
 
 class HappyWorkerWithCron < HappyWorker